chore: fix imports, move fnc to utils
