+++ packages/apps/human-app/frontend/src/modules/signup/operator/hooks/use-get-stacked-amount.ts
@@ -1,19 +1,7 @@
 import { useQuery } from '@tanstack/react-query';
-import { ethers } from 'ethers';
-import { t } from 'i18next';
 import { stakingGetStakedTokens } from '@/modules/smart-contracts/Staking/staking-get-staked-tokens';
 import { useConnectedWallet } from '@/shared/contexts/wallet-connect';
 import { getContractAddress } from '@/modules/smart-contracts/get-contract-address';
-
-export const stakedAmountFormatter = (amount: bigint) => {
-  const amountAsString = ethers.formatEther(amount);
-
-  if (amountAsString.split('.')[1] === '0') {
-    // decimals part should be omitted
-    return `${amountAsString.replace('.0', '')} ${t('inputMasks.humanCurrencySuffix')}`;
-  }
-  return `${ethers.formatEther(amount)} ${t('inputMasks.humanCurrencySuffix')}`;
-};
 
 export function useGetStakedAmount() {
   const {
